Add Listingvalidator and AdvertisementValidator

+++ components/CreateListingForm.tsx
@@ -13,6 +13,7 @@ import ListItemText from "@material-ui/core/ListItemText";
 import Select from "@material-ui/core/Select";
 import Checkbox from "@material-ui/core/Checkbox";
 import { Listing } from "./Types";
+import validateListing from "./SignUpComponents/ListingValidator";
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -130,46 +131,51 @@ const CreateListingForm = ({
               owner: state.currentUser,
               categories: selectedCategories,
             };
-            if (initialListing) {
-              dispatch({
-                type: "REMOVE_LISTING",
-                payload: initialListing.id,
-              });
-            }
-            if (state.usingDB) {
+            if (validateListing(newListing)){
+                
+                
               if (initialListing) {
-                await fetch("/api/listings", {
-                  method: "DELETE",
+                dispatch({
+                  type: "REMOVE_LISTING",
+                  payload: initialListing.id,
+                });
+              }
+              if (state.usingDB) {
+                if (initialListing) {
+                  await fetch("/api/listings", {
+                    method: "DELETE",
+                    headers: {
+                      "Content-Type": "application/json",
+                    },
+                    body: JSON.stringify({ id: initialListing.id }),
+                  });
+                }
+                const response = await fetch("/api/listings", {
+                  method: "POST",
                   headers: {
                     "Content-Type": "application/json",
                   },
-                  body: JSON.stringify({ id: initialListing.id }),
-                });
-              }
-              const response = await fetch("/api/listings", {
-                method: "POST",
-                headers: {
-                  "Content-Type": "application/json",
-                },
-                body: JSON.stringify(newListing),
-              }).then((response) => response.json());
-              if (response) {
+                  body: JSON.stringify(newListing),
+                }).then((response) => response.json());
+                if (response) {
+                  dispatch({
+                    type: "ADD_LISTING",
+                    payload: { id: response.id, ...newListing },
+                  });
+                  router.push("/");
+                }
+              } else {
                 dispatch({
                   type: "ADD_LISTING",
-                  payload: { id: response.id, ...newListing },
+                  payload: {
+                    id: Math.floor(Math.random() * Math.floor(10000000)),
+                    ...newListing,
+                  },
                 });
                 router.push("/");
               }
-            } else {
-              dispatch({
-                type: "ADD_LISTING",
-                payload: {
-                  id: Math.floor(Math.random() * Math.floor(10000000)),
-                  ...newListing,
-                },
-              });
-              router.push("/");
             }
+
           }}
           variant="contained"
           color="primary"
